+++ astropy/modeling/functional_models.py
@@ -4,6 +4,7 @@
 
 from __future__ import (absolute_import, unicode_literals, division,
                         print_function)
+import warnings
 import numpy as np
 from .core import (Fittable1DModel, Fittable2DModel, Model,
                    ModelDefinitionError, custom_model)
@@ -12,14 +13,15 @@ from .utils import ellipse_extent
 from ..utils import deprecated
 from ..extern import six
 from .utils import get_inputs_and_params
+from ..utils.exceptions import AstropyDeprecationWarning
 
 
 __all__ = ['AiryDisk2D', 'Moffat1D', 'Moffat2D', 'Box1D', 'Box2D', 'Const1D',
            'Const2D', 'Ellipse2D', 'Disk2D', 'Gaussian1D',
            'GaussianAbsorption1D', 'Gaussian2D', 'Linear1D', 'Lorentz1D',
-           'MexicanHat1D', 'MexicanHat2D', 'Redshift', 'Scale', 'Sersic1D',
-           'Sersic2D', 'Shift', 'Sine1D', 'Trapezoid1D', 'TrapezoidDisk2D',
-           'Ring2D', 'custom_model_1d', 'Voigt1D']
+           'MexicanHat1D', 'MexicanHat2D', 'RedshiftScaleFactor', 'Redshift',
+           'Scale', 'Sersic1D', 'Sersic2D', 'Shift', 'Sine1D', 'Trapezoid1D',
+           'TrapezoidDisk2D', 'Ring2D', 'custom_model_1d', 'Voigt1D']
 
 
 class Gaussian1D(Fittable1DModel):
@@ -537,44 +539,52 @@ class Scale(Model):
         return factor * x
 
 
-class Redshift(Fittable1DModel):
+class RedshiftScaleFactor(Fittable1DModel):
     """
-    One dimensional redshift model.
+    One dimensional redshift scale factor model.
 
     Parameters
     ----------
-    z : float or a list of floats
-        Redshift value(s).
+    z : float
+        Redshift value.
 
     Notes
     -----
     Model formula:
 
-        .. math:: \\lambda_{obs} = (1 + z) \\lambda_{rest}
-
+        .. math:: f(x) = x (1 + z)
     """
 
     z = Parameter(description='redshift', default=0)
 
     @staticmethod
     def evaluate(x, z):
-        """One dimensional Redshift model function"""
+        """One dimensional RedshiftScaleFactor model function"""
 
         return (1 + z) * x
 
     @staticmethod
     def fit_deriv(x, z):
-        """One dimensional Redshift model derivative"""
+        """One dimensional RedshiftScaleFactor model derivative"""
+
         d_z = x
         return [d_z]
 
     @property
     def inverse(self):
-        """Inverse Redshift model"""
+        """Inverse RedshiftScaleFactor model"""
 
         inv = self.copy()
         inv.z = 1.0 / (1.0 + self.z) - 1.0
         return inv
+
+
+class Redshift(RedshiftScaleFactor):
+    def __init__(self, *args):
+        warnings.warn('The "Redshift" class is now deprecated -- use the '
+                      '"RedshiftScaleFactor" class instead.',
+                      AstropyDeprecationWarning)
+        super(Redshift, self).__init__(*args)
 
 
 class Sersic1D(Fittable1DModel):